ocr_core/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pymupdf>=1.23.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
PyYAML>=6.0
httpx>=0.27.0
